backend/requirements.txt
django
djangorestframework
django-cors-headers
pandas
gunicorn
requests
